JBTM-1488 Minishift instructions to run quickstart under minishift
