+++ rts/lra/.gitignore
@@ -1,0 +1,3 @@
+*swarm.jar
+narayana-full-5.6.5.Final-SNAPSHOT/
+narayana-full-5.6.5.Final-SNAPSHOT-bin.zip
