+++ rts/lra/hotel-service/pom.xml
@@ -1,0 +1,127 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.jboss.narayana.quickstart.rts.lra</groupId>
+        <artifactId>lra-quickstarts</artifactId>
+        <version>5.6.5.Final-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <artifactId>hotel-service</artifactId>
+    <name>LRA Quickstart hotel-service</name>
+    <description>LRA Quickstart</description>
+    <packaging>war</packaging>
+
+    <properties>
+        <failOnMissingWebXml>false</failOnMissingWebXml>
+        <swarm.http.port>8081</swarm.http.port>
+        <lra.http.host>localhost</lra.http.host>
+        <lra.http.port>8080</lra.http.port>
+        <service.http.port>${swarm.http.port}</service.http.port>
+
+        <version.wildfly-swarm>2017.8.1</version.wildfly-swarm>
+        <version.resteasy-client>3.1.3.Final</version.resteasy-client>
+        <version.json.api>1.1</version.json.api>
+        <version.jaxrs.api>1.0.0.Final</version.jaxrs.api>
+        <version.jackson>2.8.9</version.jackson>
+
+        <version.jboss-interceptors>1.0.1.Final</version.jboss-interceptors>
+        <version.jaxrs-api>2.0</version.jaxrs-api>
+        <version.cdi-api>1.2</version.cdi-api>
+        <version.junit>4.12</version.junit>
+        <version.arquillian>1.1.12.Final</version.arquillian>
+
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+        <test.logs.to.file>false</test.logs.to.file>
+
+    </properties>
+
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>org.wildfly.swarm</groupId>
+                <artifactId>bom-all</artifactId>
+                <version>${version.wildfly-swarm}</version>
+                <scope>import</scope>
+                <type>pom</type>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+    <build>
+        <finalName>lra-test</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.wildfly.swarm</groupId>
+                <artifactId>wildfly-swarm-plugin</artifactId>
+                <version>${version.wildfly-swarm}</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>package</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <configuration>
+                    <environment>
+                        <swarm.http.port>${swarm.http.port}</swarm.http.port>
+                    </environment>
+                    <properties>
+                        <lra.http.host>${lra.http.host}</lra.http.host>
+                        <lra.http.port>${lra.http.port}</lra.http.port>
+                        <service.http.port>${service.http.port}</service.http.port>
+                    </properties>
+                </configuration>
+            </plugin>
+<!--
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-dependency-plugin</artifactId>
+        <version>3.0.1</version>
+        <configuration>
+          <artifactItems>
+            <artifactItem>
+              <groupId>org.jboss.narayana.rts</groupId>
+              <artifactId>lra-coordinator</artifactId>
+              <version>${project.version}</version>
+            </artifactItem>
+          </artifactItems>
+        </configuration>
+      </plugin>
+-->
+        </plugins>
+    </build>
+
+    <dependencies>
+        <!-- WildFly Swarm Fractions -->
+        <dependency>
+            <groupId>org.wildfly.swarm</groupId>
+            <artifactId>microprofile</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.wildfly.swarm</groupId>
+            <artifactId>transactions</artifactId>
+        </dependency>
+
+         <!-- LRA JAXRS filters -->
+        <dependency>
+            <groupId>org.jboss.narayana.rts</groupId>
+            <artifactId>lra-filters</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <!-- JAXRS 2 Client API implementation -->
+        <dependency>
+            <groupId>org.jboss.resteasy</groupId>
+            <artifactId>resteasy-client</artifactId>
+            <version>${version.resteasy-client}</version>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+</project>
