+++ rts/lra/lra-test/.gitignore
@@ -1,0 +1,2 @@
+narayana-full-5.6.5.Final-SNAPSHOT-bin.zip
+narayana-full-5.6.5.Final-SNAPSHOT/
