--- rts/lra/src/main/java/participant/demo/BookingCallback.java
@@ -1,47 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2017, Red Hat, Inc., and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package participant.demo;
-
-import participant.model.Booking;
-
-import javax.ws.rs.client.InvocationCallback;
-import java.util.concurrent.CompletableFuture;
-
-public class BookingCallback implements InvocationCallback<Booking> {
-
-    private final CompletableFuture<Booking> future = new CompletableFuture<>();
-
-    CompletableFuture<Booking> getCompletableFuture() {
-        return future;
-    }
-
-    @Override
-    public void completed(Booking booking) {
-        future.complete(booking);
-    }
-
-    @Override
-    public void failed(Throwable t) {
-        future.completeExceptionally(t);
-    }
-}
-
